requirements.txt
aiofiles==25.1.0
aiogram==3.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asyncpg==0.31.0
attrs==25.4.0
certifi==2026.2.25
click==8.3.1
distro==1.9.0
fastapi==0.134.0
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
openai==2.24.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.47
sqlmodel==0.0.37
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.41.0
yarl==1.22.0
